requirements.txt
pymongo
pandas 
numpy
python-dotenv
seaborn
requests
matplotlib
jupyter
scikit-learn
dataclasses
flask